UBI: make the control character device non-seekable
This patch makes the UBI control device (/dev/ubi_ctrl) non-seekable.
The seek operation does is not applicable to this file, so it is
cleaner to explicitly return error (which the added 'no_llseek()')
does than trying to change the position (which the removed
'default_llseek()' does).

This is an API break, but the only known user of this interface is
mtd-utils which does not need the seeking functionality. And any app
which relies on this is broken, but I'm not aware of such apps.

Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>